requirements.txt
fastapi
uvicorn
tweepy
python-dotenv
pydantic